added accounts, transactions, categories, and account management functions

+++ src/infrastructure/database/postgresql/models/accounts.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import DateTime, ForeignKey, Integer, Numeric, String, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
@@ -17,7 +18,9 @@ class Account(Base):
     name: Mapped[str] = mapped_column(String, nullable=False)
     type: Mapped[str] = mapped_column(String, nullable=False)
     currency: Mapped[str] = mapped_column(String(3), nullable=False, default="RUB")
-    balance: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False, default=0.0)
+    balance: Mapped[Decimal] = mapped_column(
+        Numeric(12, 2), nullable=False, default=0.0
+    )
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=func.now())
 
     user: Mapped["User"] = relationship(
